# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MatchApply API
  version: 1.0.0
  description: |
    API de MatchApply: autenticación con verificación por email (MFA), procesamiento
    de CVs con un modelo blindado (solo usa datos del archivo recibido) y paywall
    free/pro decidido **siempre** en el servidor.

    ## Autenticación
    Cookies httpOnly (`ma_at` de acceso, `ma_rt` de refresh rotativo). El cliente
    llama con `credentials: 'include'`. Nunca se guardan tokens en localStorage.

    ## Errores
    Formato uniforme: `{ "error": { "code": "...", "message": "..." } }`
servers:
  - url: https://api.matchapply.com
  - url: http://localhost:8080
...
paths:
...
  /auth/signup:
    post:
      tags: [auth]
      summary: Crea la cuenta SIN verificar y envía el código de 6 dígitos por email
      description: |
        Valida las 5 reglas de contraseña en el servidor (la validación del cliente es solo UX).
        **El código nunca vuelve en la respuesta.** Rate limit: 5 registros por IP cada 15 min.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, passwordConfirm]
              properties:
                name: { type: string, minLength: 2, maxLength: 80 }
                email: { type: string, format: email }
                password: { type: string, minLength: 8, maxLength: 128 }
                passwordConfirm: { type: string }
                isDiscoverable: { type: boolean, default: false }
      responses:
        '202':
          description: Pendiente de verificación
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: pending_verification }
                  email: { type: string }
                  expiresInMinutes: { type: integer, example: 15 }
                  resendCooldownSeconds: { type: integer, example: 30 }
        '400':
          description: weak_password | password_mismatch | invalid_payload
        '429': { description: signup_rate_limited }
  /auth/verify:
    post:
      tags: [auth]
      summary: Valida el código, activa la cuenta y abre sesión
      description: 15 minutos de validez, 5 intentos, comparación en tiempo constante.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: '^\d{6}$' }
      responses:
        '200':
          description: Cuenta verificada; se setean las cookies de sesión
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '400':
          description: invalid_code | code_expired | code_too_many
  /auth/resend:
    post:
      tags: [auth]
      summary: Reenvía el código (cooldown de 30s)
      responses:
        '202': { description: Enviado (respuesta uniforme aunque el mail no exista) }
        '429': { description: resend_cooldown }
  /auth/login:
    post:
      tags: [auth]
      summary: Inicia sesión
      description: |
        Mismo mensaje para usuario inexistente y contraseña incorrecta (anti-enumeración).
        Bloqueo temporal de la cuenta a los 8 fallos. Rate limit: 10 por IP cada 10 min.
      responses:
        '200': { description: Sesión iniciada }
        '401': { description: invalid_credentials }
        '403': { description: not_verified (se reenvía el código automáticamente) }
        '429': { description: account_locked | login_rate_limited }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rota el refresh token y renueva el access token
      responses:
        '200': { description: Renovado }
        '401': { description: session_expired }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ma_at
  schemas:
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        tier: { type: string, enum: [free, pro] }
        isVerified: { type: boolean }
...
security:
  - cookieAuth: []